models/requirement.txt
torch
torchvision
pillow==6.2.0
nltk
editdistance
pandas
numpy
matplotlib